Backend/requirements.txt
Flask
pydantic
google-generativeai
python-dotenv
Flask-Cors